\name{biocValues2Matrix}
\alias{biocValues2Matrix}
\title{generate three matrices for the data over all the files}
\usage{
  biocValues2Matrix(useDir, filePattern, minIndex = 2)
}
\arguments{
  \item{useDir}{which directory to read the files from}

  \item{filePattern}{the pattern to use to list the files}

  \item{minIndex}{what is the minimum index of the file in
  chronological order}
}
\description{
  generate three matrices for the data over all the files
}

