\name{checkBiocPackages}
\alias{checkBiocPackages}
\title{checks the content of the data on the SVN server provided}
\usage{
  checkBiocPackages(packageList, baseURL,
    usePWD = "readonly:readonly")
}
\arguments{
  \item{packageList}{the list of packages to check}

  \item{baseURL}{the url of the svn server to work with}

  \item{usePWD}{the svn password to use to get access to
  files}
}
\description{
  checks the content of the data on the SVN server provided
}

