\name{checkBiocVersion}
\alias{checkBiocVersion}
\title{checks the version of the package on the server}
\usage{
  checkBiocVersion(inURL, usePWD = "readonly:readonly")
}
\arguments{
  \item{inURL}{full url to the package}

  \item{usePWD}{the username and password to use}
}
\description{
  checks the version of the package on the server
}

