\name{getPackageFiles}
\alias{getPackageFiles}
\title{gets the list of packages contained in a specified directory}
\usage{
  getPackageFiles(checkDir)
}
\arguments{
  \item{checkDir}{the directory containing the packages}
}
\value{
  the returned values in the list are: \itemize{
  \item{files: }{the tar files} \item{package: }{the
  package names} \item{version: }{the package versions} }
}
\description{
  gets the list of packages contained in a specified
  directory
}

