\name{hasRmdVignette}
\alias{hasRmdVignette}
\title{searches the file listing for an "Rmd" vignette}
\usage{
  hasRmdVignette(inList,
    grepStr = c(".Rmd", ".rmd", ".RMD"))
}
\arguments{
  \item{fileList}{the list of files to grep over}

  \item{grepStr}{a list of strings to search for presence
  of in file listing}
}
\description{
  searches the file listing for an "Rmd" vignette
}

